Use BigInt for fixture numericId (32-bit overflow)
toNumericId() takes 8 hex digits = up to 4.29e9, which overflows Postgres
INTEGER (max ~2.1e9) — the refresh-fixtures cron 500'd with "integer out of
range" on ids like 4271696926. Switch FixtureCache/FixtureResult numericId to
BigInt and convert at the DB boundary. Prod columns already altered to BIGINT.

+++ prisma/manual/2026-06-09-fixtures-cache.sql
@@ -10,7 +10,7 @@ CREATE INDEX IF NOT EXISTS "bet_selections_fixture_id_idx" ON "bet_selections" (
 
 -- 2. Cache of normalized fixtures (live + upcoming), refreshed by the cron
 CREATE TABLE IF NOT EXISTS "fixtures_cache" (
-  "numeric_id"    INTEGER      NOT NULL,
+  "numeric_id"    BIGINT       NOT NULL,
   "event_id"      TEXT         NOT NULL,
   "sport_key"     TEXT         NOT NULL,
   "commence_time" TIMESTAMP(3) NOT NULL,
